create service interfaces for certification, adapt default config
